chore: refactor code part 2
